+++ packages/jest-remirror/src/patches.d.ts
@@ -1,3 +1,0 @@
-declare module '@testing-library/react/pure' {
-  export * from '@testing-library/react';
-}